FinalProject/requirements.txt
alembic==0.8.8
asgi-redis==1.0.0
asgiref==1.0.0
attrs==16.3.0
autobahn==17.10.1
Automat==0.5.0
awsebcli==3.12.1
blessed==1.14.2
boto==2.42.0
boto3==1.4.7
botocore==1.7.43
cement==2.8.2
channels==1.0.3
colorama==0.3.7
constantly==15.1.0
daphne==1.0.3
dataset==0.6.4
defusedxml==0.5.0
Django==1.9.6
django-allauth==0.25.2
django-amazon-ses==0.3.0
django-appconf==1.0.2
django-bootstrap-form==3.3
django-bootstrap3==9.1.0
django-bugtracker==1.1.0
django-extensions==1.9.6
django-graphos==0.3.3
django-gravatar==0.1.0
django-gravatar2==1.4.2
django-smtp-ssl==1.0
django-storages==1.1.8
django-storages-redux==1.3.2
django-widget-tweaks==1.4.1
djangorestframework==3.5.3
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.14
gunicorn==19.6.0
hyperlink==17.3.1
icalendar==4.0.0
incremental==16.10.1
jmespath==0.9.3
Mako==1.0.4
MarkupSafe==0.23
msgpack-python==0.4.8
normality==0.2.4
oauthlib==2.0.6
olefile==0.44
pathspec==0.5.0
Pillow==4.3.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
python-editor==1.0.1
python3-openid==3.1.0
pytz==2017.2
PyYAML==3.12
redis==2.10.5
requests==2.9.1
requests-oauthlib==0.8.0
s3transfer==0.1.11
semantic-version==2.5.0
six==1.11.0
solid-i18n==1.4.1
SQLAlchemy==1.0.15
tabulate==0.7.5
tellme==0.2.6
termcolor==1.1.0
texttable==0.8.4
Twisted==17.1.0
txaio==2.6.1
txredisapi==1.4.4
wcwidth==0.1.7
websocket-client==0.44.0
XlsxWriter==0.8.9
zope.interface==4.3.3
